chat_summarizer/requirements.txt
requests>=2.31.0
huggingface_hub>=0.20.0
python-dotenv>=1.0.0
